Refactor unique charts to inherit from base
