+++ lib/apexcharts/charts/base.rb
@@ -2,8 +2,21 @@ module Apexcharts
   class BaseChart
     attr_reader :options, :series
 
+    def initialize data, options={}
+      @series = sanitize_data(data)
+      @options = Utils::Hash.deep_merge(
+                   build_options(x_sample, options),
+                   Utils::Hash.camelize_keys(
+                     {**@series, chart: {type: chart_type}}.compact
+                   )
+                 )
+    end
+
     def render
       Apexcharts::Renderer.render_default(options)
+    end
+
+    def chart_type
     end
 
   protected
