+++ lib/apexcharts/charts/polar.rb
@@ -1,17 +1,5 @@
 module Apexcharts
   class PolarChart < BaseChart
-    def initialize data, options={}
-      @series = sanitize_data(data)
-      @options = Utils::Hash.deep_merge(
-                   build_options(x_sample, options),
-                   Utils::Hash.camelize_keys(
-                     {**@series, chart: {type: chart_type}}.compact
-                   )
-                 )
-    end
-
-    def chart_type
-    end
 
   protected
