+++ lib/apexcharts/charts/radar.rb
@@ -1,12 +1,15 @@
 # frozen_string_literal: true
 
 module Apexcharts
-  class RadarChart < CartesianChart
+  class RadarChart < BaseChart
     def chart_type
       'radar'
     end
 
-    def mixed_series
+  private
+
+    def sanitize_data(data)
+      Apexcharts::CartesianSeries.new(data).sanitized
     end
   end
 end
