+++ spec/charts/cartesian_spec.rb
@@ -11,15 +11,6 @@ RSpec.describe Apexcharts::CartesianChart do
     chart = described_class.new(bindings, data, options)
     expect(chart.chart_type).to eq(nil)
     expect(chart.more_options).to eq({})
-    expect(chart.mixed_series).to eq(
-      [
-        {data: [
-          {x: 100, y: 1},
-          {x: 200, y: 2}
-        ],
-        type: nil}
-      ]
-    )
   end
 end
